requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.0.14
mssql-django==1.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
idna==3.11
pymongo==4.16.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
